Partner Spend y-m + media _source
The query will provide spend amounts for each partner by month (YYYY-MM) and media source. This query can be used as a basis for tracking spend of all RAP partners.

Spend is determined by KPI percentage and total number of profiles generated. This should not be used to calculate spend for Strategic partners.